requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==1.7.1
pytest==8.3.5
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
